Implemented user access with query params

+++ UserManagement/src/main/scala/Clea.scala
@@ -1,6 +1,8 @@
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
+import akka.http.scaladsl.model.{HttpResponse, StatusCodes}
 import akka.http.scaladsl.server.Directives._
+import akka.http.scaladsl.unmarshalling.{Unmarshaller, _}
 import akka.stream.ActorMaterializer
 import com.google.gson.Gson
 import com.typesafe.config.ConfigFactory
@@ -8,6 +10,7 @@ import com.typesafe.scalalogging.Logger
 import org.bson.types.ObjectId
 import pdi.jwt.{Jwt, JwtAlgorithm, JwtClaim, JwtHeader}
 
+import scala.concurrent.Future
 import scala.io.StdIn
 
 /**
@@ -66,6 +69,31 @@ case class BookRecord(userId: String,
 
 case class JWTPayload(iat: Long, exp: Long, sub: String, role: String)
 
+trait CsvParameters {
+  implicit def csvSeqParamMarshaller: FromStringUnmarshaller[Seq[String]] =
+    Unmarshaller(ex => s => Future.successful(s.split(",")))
+
+  implicit def csvListParamMarshaller: FromStringUnmarshaller[List[String]] =
+    Unmarshaller(ex => s => Future.successful(s.split(",").toList))
+}
+
+final object CsvParameters extends CsvParameters
+
+object UserSearchCriteria {
+  def apply(usersIds: Option[List[String]], region: Option[String]) =
+    new UserSearchCriteria(usersIds.getOrElse(List()), region.getOrElse(""))
+}
+
+case class UserSearchCriteria(userIds: List[String] = List(), region: String = "")
+
+object BookRecordSearchCriteria {
+  def apply(dateFrom: Option[Long], dateTo: Option[Long], usersIds: Option[List[String]], region: Option[String]) =
+    new BookRecordSearchCriteria(dateFrom.getOrElse(-1), dateTo.getOrElse(-1), usersIds.getOrElse(List()), region.getOrElse(""))
+}
+
+case class BookRecordSearchCriteria(dateFrom: Long = -1, dateTo: Long = -1, userIds: List[String] = List(), region: String = "")
+
+
 class Clea
 
 object Clea {
@@ -81,7 +109,7 @@ object Clea {
     var claim = JwtClaim()
     claim = claim +("iat", System.currentTimeMillis())
     claim = claim +("exp", System.currentTimeMillis() + 86400)
-    claim = claim +("sub", "admin")
+    claim = claim +("sub", user.username)
     claim = claim +("role", user.role)
 
     Jwt.encode(header, claim, secret_key)
@@ -96,7 +124,7 @@ object Clea {
     UserManagement.setup
 
     val route = {
-      var role = "none"
+      var payload: JWTPayload = null
       var token = ""
 
       pathSingleSlash {
@@ -107,9 +135,7 @@ object Clea {
             post {
               entity(as[String]) { loginSpecJson => {
                 logger.info("Commencing login")
-
                 val loginSpec = new Gson().fromJson(loginSpecJson, classOf[LoginSpec])
-
                 val user = UserManagement.login(loginSpec)
 
                 user match {
@@ -127,9 +153,12 @@ object Clea {
           logger.info("About to auth")
           if (authHeader.isPresent) {
             token = authHeader.get().value()
+
             logger.info(s"Token: $token")
-            logger.info(s"Is blacklisted: ${!UserManagement.isTokenBlacklisted(token)}")
+            logger.info(s"Is blacklisted: ${UserManagement.isTokenBlacklisted(token)}")
             logger.info(s"Is valid: ${Jwt.isValid(token, secret_key, Seq(JwtAlgorithm.HS512))}")
+
+            payload = new Gson().fromJson(Jwt.decode(token, secret_key, Seq(JwtAlgorithm.HS512)).get, classOf[JWTPayload])
             !UserManagement.isTokenBlacklisted(token) && Jwt.isValid(token, secret_key, Seq(JwtAlgorithm.HS512))
           } else false
         }) {
@@ -144,14 +173,25 @@ object Clea {
                 post {
                   entity(as[String]) {
                     userSpecJson => {
-                      val userSpec = new Gson().fromJson(userSpecJson, classOf[UserSpec])
-                      complete(s"New user created by role $role")
+                      if(payload.role.equalsIgnoreCase("admin")) {
+                        val userSpec = new Gson().fromJson(userSpecJson, classOf[UserSpec])
+                        val newUser = UserManagement.createUser(userSpec)
+                        complete(new Gson().toJson(UserExposed(newUser)))
+                      } else {
+                        complete(HttpResponse(StatusCodes.Unauthorized))
+                      }
                     }
                   }
                 } ~
                   get {
-                    val allUsers = UserManagement.getAllUsers().map(UserExposed.apply(_))
-                    complete(new Gson().toJson(allUsers.toArray))
+                    import CsvParameters._
+
+                    parameters('users.as[List[String]].?, 'region.as[String].?) {
+                      (users, region) => {
+                        val allUsers = UserManagement.getUsers(UserSearchCriteria(users, region)).map(UserExposed.apply(_))
+                        complete(new Gson().toJson(allUsers.toArray))
+                      }
+                    }
                   }
               } ~
                 pathPrefix("me") {
@@ -167,23 +207,46 @@ object Clea {
                   username => {
                     pathEnd {
                       get {
-                        val user = UserExposed(UserManagement.getByUsername(username))
-                        complete(new Gson().toJson(user))
+                        payload.role match {
+                          case "admin" => {
+                            complete(new Gson().toJson(UserExposed(UserManagement.getByUsername(username))))
+                          }
+                          case "manager" => {
+                            val user = UserManagement.getByUsername(username)
+                            val manager = UserManagement.getByUsername(payload.sub)
+
+                            if(manager.region.equalsIgnoreCase(user.region))
+                              complete(new Gson().toJson(UserExposed(UserManagement.getByUsername(username))))
+                            else
+                              complete(HttpResponse(StatusCodes.Unauthorized))
+                          }
+                          case _ => {
+                            if(payload.sub.equalsIgnoreCase(username))
+                              complete(new Gson().toJson(UserExposed(UserManagement.getByUsername(username))))
+                            else
+                              complete(HttpResponse(StatusCodes.Unauthorized))
+                          }
+                        }
                       } ~
                         delete {
-                          UserManagement.deleteUser(username)
-                          complete(s"User $username deleted")
+                          if(payload.role.equalsIgnoreCase("admin")) {
+                            UserManagement.deleteUser(username)
+                            complete(s"User $username deleted")
+                          } else
+                            complete(HttpResponse(StatusCodes.Unauthorized))
                         } ~
                         patch {
                           entity(as[String]) {
                             userSpecJson => {
-                              val userSpec = new Gson().fromJson(userSpecJson, classOf[UserSpec])
-                              val updatedUser = UserManagement.updateUser(userSpec)
-                              complete(new Gson().toJson(UserExposed(updatedUser)))
+                              if (payload.sub.equalsIgnoreCase("admin")) {
+                                val userSpec = new Gson().fromJson(userSpecJson, classOf[UserSpec])
+                                complete(new Gson().toJson(UserExposed(UserManagement.updateUser(userSpec))))
+                              } else
+                                complete(HttpResponse(StatusCodes.Unauthorized))
                             }
                           }
                         }
-                      }
+                    }
                   }
                 }
             } ~
